Commit some in-progress stuff.
In the midst of disassembling a huge function in logic.asm, i decided to
disassembly the entirety of segment 7. I'm obviously not going to be
finished with either of these anytime soon, so i figured i'd commit what
i have.

The logic function seems to coordinate monster movement - for each
monster in the monster list it picks a direction and calls the move tile
function, which is in segment 7. Segment 7 also contains some code
dealing with mouse movement.

I copied some of the constants from logic.asm to seg7.asm. Need to
move them to a dedicated file.

Also, desperately need to figure out what to do about linking.