make calls less hardcoded in logic.asm
For calls to functions within the same segment, we can use the function
symbols in the call instruction instead of hardcoding the address.
(Why are these even far calls? They could all be near calls afaik.)

We can't do it for calls to functions in other segments because they're
in a different compilation unit so we don't know their addresses. And we
can't do it for calls to imported symbols.

And we still can't do anything about the relocation chain clogging up
the segment specifier.

But hey, it's a start!