clean up makefile slightly
bin files already have a dependency on the corresponding asm file and
the Makefile due to the generic rule. Therefore we don't need to specify
it manually.

    %.bin: %.asm fixmov.sh Makefile

Tweak the dependency lines so they only list additional dependencies
(that is, included files).

Note that we can remove data.bin entirely, since it has no additional
dependencies.