fix variables.asm dependency
Although it's true that variables.asm depends (indirectly) on data.asm,
genvars.sh actually derives it from data.map, which is generated as a
by-product of generating data.bin. Because this dependency link was
missing, it was possible to run `make variables.asm` after editing
data.asm and wind up with a stale file.

Fix this by declaring a dependency on data.bin instead. *Don't* depend
on data.map, because make doesn't actually know how to build it. There's
a way to tell make about rules that build multiple targets, but that
could get complicated and i'm not sure if it's portable. Instead, just
declare the dependency on the thing make knows about, data.bin.