seg2.asm: remove default function names from comments
tools/call won't add a function name to a call if there's already one
there, so having the autogenerated names here is interfering with the
ability to be able to re-run tools/call on a file to add new names.

I've already modified tools/call to not write these in the first place.
Just need to clean up seg2.asm.