seg2.asm: comment some of the viewport code
Name the UpdateChip and ScrollViewport functions, add some arguments and
locals, and add a bunch of comments.

The comments were mostly for myself as i tried to figure what was going
on with viewport calculations and how UnusedOffsetX and UnusedOffsetY
fit into it, so they are a little terse and might not make much
sense. I'll probably go back and clean them up sometime.