fix ending animation, mostly
calculate the right number of frames to grow so that chip fills up the
entire viewport. increase the growth rate from 8 pixels/frame to 16
pixels a frame accordingly, so that the number of frames fits into a
byte.

the chipend graphic is still hardcoded at 0x120 by 0x120 pixels because
we can't change its size.