deconflate Monster struct
We were reusing Monster for both the monster list and the slip list
because they happened to be the same size and have mostly the same
layout. However, they had different semantics for the slipping flag,
which was very confusing (on the monster list it means the monster is
slipping; on the slip list it means the slipper is a block). Fix that by
introducing a separate Slipper type for the slip list.