tools/find_reloc_runs.py: new tool
Analyzes the order of the relocation patch lists and finds spans of
ascending or descending addresses.

I haven't found any obvious pattern to the way either relocations or
patches are ordered ordered, so the linker is going to need some sort of
config file that specifies the order. This tool is the first step
towards generating that config file.