tools/linker: progress on symbol resolution
Turns out the fixup code was incomplete: we weren't even trying to
properly resolve symbols or put patches in the right chain yet, we were
just building a giant chain of every single patch.

So before trying to hack support in for putting relocation chains in
the right order, let's try to at least get them in the wrong order first.

This commit is a step in that direction.