tools/link: split Segment into two types
for bootstrapping purposes it would be nice to be able to link to symbol
in segments that don't exist (that is, aren't among the input files).
that should be fine - we only need to know a segment's number to
reference a symbol in it -  but the way the types are laid out
currently, Symbol points to Segment, which has a bunch of other fields.

solution: split Segment into two types, Segment and SegmentInfo. Segment
acts as a lightweight identifier; it only has an index field. (could
possibly make it a value type but the ability to lazily assign numbers
might come in handy later on). SegmentInfo holds all the bookkeeping
data for when we actually need to link code inside the segment.