Change linker output from foo.linked.bin to foo.bin
I was initially worried that make would get confused if we had .bin
files being generated by two different routes (asm->bin for data.asm,
asm->obj->bin for the code segments), thus the extra extension in the
filename; but, as it turned out, the linker-generated ones aren't
actually part of the dependency graph so it's fine.

And data.asm will eventually go through the same path as the others so
it shouldn't be an issue in the future either.