Makefile: streamline link step
Combine the link step with the final assembly of chips.asm.
This is how linking would work normally anyway (read a bunch of
object files, produce an exe), it just happens that our linker
is split across two programs and produces a bunch of intermediate
.bin files.

This allows us to get rid of the link.stamp hack.