*.asm: jnc => jnb
jnc = jump if carry flag not set
jnb = jump if not below

same deal as the previous commit. both are an unsigned greater-or-equal
check; use the better name.  could also use jae (jump if above or equal)
but we're already using jna instead of jbe so go with the more
consistent choice.