Simplify extern declarations
NASM's manual claims that, quote, "if a variable is declared both
`GLOBAL' and `EXTERN' [...] it will be treated as `GLOBAL'". Which is
half-true. If you declare it global and then extern, that works -
the extern is ignored. But if you declare it extern and then global you
get an error. So that's useless.

Fortunately there's a third option: if you declare something EXTERN and
then later give it a definition, it will be upgraded to GLOBAL. And of
course multiple EXTERNs are fine. Which gives us exactly what we need.
We can declare all functions EXTERN up front and not even worry about
GLOBAL.